Moved some AMV stuff into a Volume base class
